+++ src/core/core.cpp
@@ -17,7 +17,6 @@
 #include "core/hle/service/sm/sm.h"
 #include "core/hw/hw.h"
 #include "core/loader/loader.h"
-#include "core/memory_setup.h"
 #include "core/settings.h"
 #include "file_sys/vfs_real.h"
 #include "video_core/video_core.h"
