+++ src/core/hle/kernel/scheduler.h
@@ -8,6 +8,7 @@
 #include <vector>
 #include "common/common_types.h"
 #include "common/thread_queue_list.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/thread.h"
 
 class ARM_Interface;
