+++ src/core/hle/kernel/server_port.h
@@ -9,7 +9,7 @@
 #include <tuple>
 #include <vector>
 #include "common/common_types.h"
-#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/object.h"
 #include "core/hle/kernel/wait_object.h"
 
 namespace Kernel {
